Vertical Range backdrop filter and improved tooltip positioning
Improve positioning logic for tooltip so it is more accurate and fix the positioning for top-down (eg. shades). Backdrop filter is applied so it looks better on light colored and complex backgrounds. Set the tooltip value when updated programmatically too for users who modify the CSS to keep the tooltip always displayed.