Homey Custom Capabilities Tile
This tile supports Homey devices with Custom ENUM capabilities. It displays the current value and when tapped will display the list of available values. You can supply a `translations` JSON string in the format `{"home": "Maison"}` where the key on the left is the raw value used by homey and the value on the right is your preferred translated value.